Break up logic into multiple distinct files to separate concerns.
